including functions for energy-based and maximum-filter based keypoint detections
