+++ src/aind_registration_evaluation/metric/__init__.py
@@ -4,4 +4,5 @@
 
 from .factory import ImageMetricsFactory
 from .large_scale import LargeImageMetrics
-from .small_scale import SmallImageMetrics
+from .small_scale import (SmallImageMetrics, compute_feature_space_distances,
+                          get_pairs_from_distances)
