+++ src/aind_registration_evaluation/sample/__init__.py
@@ -5,3 +5,4 @@ areas
 # flake8: noqa: F403
 
 from .sample_points import *
+from .utils import *
