+++ src/aind_registration_evaluation/util/intersection.py
@@ -2,7 +2,7 @@
 Module to check image intersections
 """
 
-from typing import Tuple
+from typing import List, Tuple
 
 import numpy as np
 
@@ -140,3 +140,231 @@ def calculate_bounds(
     bound_2 = np.array([coord_1, coord_2])
 
     return bound_1, bound_2
+
+
+# TODO Change to adaptive non-max supression based on
+# response threshold not overlap_threshold
+def kd_non_max_suppression(
+    nd_boxes: List[List[int]], overlap_threshold: float, order_axis: str = "y"
+):
+    """
+    Applies n-dimensional non-maxima
+    suppression to a set of bounding
+    boxes/cubes using a relative
+    overlap threshold.
+
+    Parameters
+    -----------
+    boxes: List[List[int]]
+        List of boxes with where the
+        form depends on its dimensionality.
+        If it's a 2D image, boxes will be
+        [y_start, x_start, y_end, x_end] and
+        3D [z_start, y_start, x_start, z_end,
+        y_end, x_end]. Start refers to the
+        top-left corner and end to top-bottom
+        corner.
+
+    overlap_threshold: float
+        Refers to the relative overlap threshold
+        between two bounding boxes or cubes.
+
+    order_axis: str
+        Axis used to order boxes/cubes.
+        Default: 'y'
+
+    Returns
+    ----------
+    List[int]
+        Set of indices after removing overlapped
+        boxes/cubes
+    """
+    n_nd_boxes = len(nd_boxes)
+    order_axis = order_axis.lower()
+
+    if n_nd_boxes <= 1:
+        return nd_boxes
+
+    if nd_boxes.dtype.kind == "i":
+        nd_boxes = nd_boxes.astype(np.float32)
+
+    if order_axis not in ["z", "y", "x"]:
+        raise NotImplementedError("This dimension is not available")
+
+    pruned_indices = []
+    dimensionality = len(nd_boxes[0]) // 2
+
+    if dimensionality == 2 and order_axis == "z":
+        raise ValueError(
+            "Providing ordering in Z axis, but only YX available."
+        )
+
+    order_axis_pos = {"z": -3, "y": -2, "x": -1}
+
+    # grab the coordinates of the bounding boxes/cubes
+    coords = []
+    area_list = []
+    # Order is [Y X Y X] or [Z Y X Z Y X]
+    for idx_dim in range(0, dimensionality * 2, 2):
+        coords.append(nd_boxes[:, idx_dim])
+        coords.append(nd_boxes[:, idx_dim + 1])
+
+    for idx_dim in range(dimensionality):
+        # [axis]2 - [axis]1 + 1 where [axis] = [Z, Y, X] for 3D
+        area_list.append(
+            nd_boxes[:, idx_dim + dimensionality] - nd_boxes[:, idx_dim] + 1
+        )
+
+    # compute the area of the boxes/cubes and sort
+    # by the bottom-right order axis
+    if dimensionality == 2:
+        area = area_list[0] * area_list[1]
+    else:
+        area = area_list[0] * area_list[1] * area_list[2]
+
+    idxs = np.argsort(coords[order_axis_pos[order_axis]])
+
+    axis = list(range(dimensionality))
+
+    while len(idxs) > 0:
+        # grab the last index in the indexes list and add the
+        # index value to the list of picked indexes
+        last = len(idxs) - 1
+        i = idxs[last]
+        pruned_indices.append(i)
+        box_cube_area = []
+        for ax in axis:
+            # e.g., y_end, y_start
+            axis_end = np.minimum(
+                coords[ax + dimensionality][i],  # accessing end corner
+                coords[ax + dimensionality][idxs[:last]],
+            )
+
+            axis_start = np.maximum(
+                coords[ax][i],  # accessing start corner
+                coords[ax][idxs[:last]],
+            )
+
+            box_cube_area.append(np.maximum(0, axis_end - axis_start + 1))
+
+        if dimensionality == 2:
+            box_cube_area = box_cube_area[0] * box_cube_area[1]
+        else:
+            box_cube_area = (
+                box_cube_area[0] * box_cube_area[1] * box_cube_area[1]
+            )
+
+        # compute the ratio of overlap
+        overlap = box_cube_area / area[idxs[:last]]
+
+        # delete all boxes with higher overlap
+        # than the threshold
+        idxs = np.delete(
+            idxs,
+            np.concatenate(([last], np.where(overlap > overlap_threshold)[0])),
+        )
+    # return only remaining bboxs/cubes
+    return np.array(pruned_indices, dtype=int)
+
+
+def generate_corner(val, window_size, mode, img_shape, axis):
+    possible_val = val + window_size
+
+    if mode == "summation":
+        if possible_val > img_shape[axis] - 1:
+            possible_val = img_shape[axis] - 1
+
+    elif mode == "subtraction":
+        possible_val = val - window_size
+        if possible_val < 0:
+            possible_val = 0
+
+    else:
+        raise ValueError("Mode not available")
+
+    return possible_val
+
+
+def kd_compute_bboxs_cubes(
+    points: List[int], window_size: int, img_shape: Tuple[int]
+):
+    n_dims = len(img_shape)
+    img_geometries = []
+
+    if n_dims == 2:
+        img_geometries = np.array(
+            [
+                np.array(
+                    [
+                        # Start top left corner
+                        generate_corner(
+                            y_val,
+                            window_size,
+                            "subtraction",
+                            img_shape,
+                            axis=-2,
+                        ),
+                        generate_corner(
+                            x_val,
+                            window_size,
+                            "subtraction",
+                            img_shape,
+                            axis=-1,
+                        ),
+                        # End bottom right corner
+                        generate_corner(
+                            y_val, window_size, "summation", img_shape, axis=-2
+                        ),
+                        generate_corner(
+                            x_val, window_size, "summation", img_shape, axis=-1
+                        ),
+                    ]
+                )
+                for y_val, x_val in points
+            ]
+        )
+
+    elif n_dims == 3:
+        img_geometries = np.array(
+            [
+                np.array(
+                    [
+                        # Start top left corner
+                        generate_corner(
+                            z_val,
+                            window_size,
+                            "subtraction",
+                            img_shape,
+                            axis=-3,
+                        ),
+                        generate_corner(
+                            y_val,
+                            window_size,
+                            "subtraction",
+                            img_shape,
+                            axis=-2,
+                        ),
+                        generate_corner(
+                            x_val,
+                            window_size,
+                            "subtraction",
+                            img_shape,
+                            axis=-1,
+                        ),
+                        # End bottom right corner
+                        generate_corner(
+                            z_val, window_size, "summation", img_shape, axis=-3
+                        ),
+                        generate_corner(
+                            y_val, window_size, "summation", img_shape, axis=-2
+                        ),
+                        generate_corner(
+                            x_val, window_size, "summation", img_shape, axis=-1
+                        ),
+                    ]
+                )
+                for z_val, y_val, x_val in points
+            ]
+        )
+
+    return img_geometries
